Moved hud to folder
