+++ TowerDefence_UnityProject/Packages/TowerDefence/Entities/Enemies/Models/IEnemiesModel.cs
@@ -1,9 +1,10 @@
-﻿using DataBinding.BaseClasses;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
+using DataBinding.BaseClasses;
+using DataBinding.Helpers;
 
 namespace TowerDefence.Entities.Enemies.Models
 {
-    [DataBinding.Helpers.DataModel(AddToZenject = true, Shared = true)]
+    [DataModel(AddToZenject = true, Shared = true)]
     public interface IEnemiesModel : IModelBase
     {
         public IList<IEnemyObject> Enemies { get; set; }
