+++ TowerDefence_UnityProject/Packages/TowerDefence/Entities/Towers/Components/AttackVisualizers/BaseAttackVisualizer.cs
@@ -1,11 +1,11 @@
 ﻿using System.Collections.Generic;
 using System.Threading.Tasks;
 using DataBinding;
+using TowerDefence.Entities.Components;
 using TowerDefence.Entities.Enemies;
 using TowerDefence.Entities.Towers.Components.Interfaces;
 using TowerDefence.Entities.Towers.Models;
 using TowerDefence.Utility;
-using IComponent = TowerDefence.Entities.Components.IComponent;
 
 namespace TowerDefence.Entities.Towers.Components.AttackVisualizers
 {
@@ -13,23 +13,23 @@ namespace TowerDefence.Entities.Towers.Components.AttackVisualizers
     {
         protected IDamageComponent DamageComponent;
         protected ITowerObject TowerObject;
-        private readonly BindingContext bindingContext = new ();
-        
+        private readonly BindingContext bindingContext = new();
+
         public void Dispose()
         {
             if (DamageComponent is not null)
             {
                 DamageComponent.AppliedDamageToTargets -= OnTargetsDamaged;
             }
-            
+
             bindingContext.Dispose();
         }
-        
+
         public virtual Task AsyncPostInit(ITowerObject towerObject, ITowerModel towerModel)
         {
             TowerObject = towerObject;
-            bindingContext.Bind(towerModel, x=>x.Components, OnComponentsChanged);
-            return  Task.CompletedTask;
+            bindingContext.Bind(towerModel, x => x.Components, OnComponentsChanged);
+            return Task.CompletedTask;
         }
 
         private void OnComponentsChanged(IList<IComponent> obj)
@@ -38,7 +38,7 @@ namespace TowerDefence.Entities.Towers.Components.AttackVisualizers
             {
                 DamageComponent.AppliedDamageToTargets -= OnTargetsDamaged;
             }
-            
+
             if (obj.TryGetComponent(out DamageComponent))
             {
                 DamageComponent.AppliedDamageToTargets += OnTargetsDamaged;
